fix gmres vpgmres bug, fix all N NNZ size from int -> unsigned long int

+++ cudaFunction.hpp
@@ -11,7 +11,7 @@ class cuda {
     double *cu_d4, *cu_d5, *cu_d6, *cu_d7, *cu_d8, *cu_h2, *cu_h3, *cu_h4;//kskipcg N*(2*k+1) N*(2*k+1) Blocks*(2*k) Blocks*(2*k+1) Blocks*(2*k+2) Blocks*(2*k) Blocks*(2*k+1) Blocks*(2*k+2)
     double *cu_d9, *cu_h5;//kskipbicg Blocks*(2*k+1) Blocks*(2*k+1)
     double *cu_d10, *cu_h6, *cu_d11, *cu_h7;//gmres vpgmres restart*N restart*N, blocks, blocks
-    int size;
+    unsigned long int size;
     int k;
     int restart;
 
@@ -19,57 +19,57 @@ class cuda {
     times *time;
 
     cuda(times *t);
-    cuda(times *t, int size);
-    cuda(times *t, int size, int k);
-    cuda(times *t, int size, double restart);
-    cuda(times *t, int size, int k, double restart);
+    cuda(times *t, unsigned long int size);
+    cuda(times *t, unsigned long int size, int k);
+    cuda(times *t, unsigned long int size, double restart);
+    cuda(times *t, unsigned long int size, int k, double restart);
     ~cuda();
 
     //--------------------------------------
-    double* d_Malloc(int size);
-    double* d_MallocHost(int size);
+    double* d_Malloc(unsigned long int size);
+    double* d_MallocHost(unsigned long int size);
 
-    int* i_Malloc(int size);
-    int* i_MallocHost(int size);
+    int* i_Malloc(unsigned long int size);
+    int* i_MallocHost(unsigned long int size);
 
-    void H2D(double* from, double* to, int size);
-    void D2H(double* from, double* to, int size);
+    void H2D(double* from, double* to, unsigned long int size);
+    void D2H(double* from, double* to, unsigned long int size);
 
-    void H2D(int* from, int* to, int size);
-    void D2H(int* from, int* to, int size);
+    void H2D(int* from, int* to, unsigned long int size);
+    void D2H(int* from, int* to, unsigned long int size);
 
     void Free(void* ptr);
 
     void FreeHost(void* ptr);
 
-    void Memset(double *ptr, double val, int size);
+    void Memset(double *ptr, double val, unsigned long int size);
 
-    void Memset(int *ptr, int val, int size);
+    void Memset(int *ptr, int val, unsigned long int size);
 
     void Reset();
 
     //--------------------------------------
-    void CSR2CSC(double *dCSRval, int *dCSRcol, int *dCSRptr, double *CSCval, int *CSCrow, int *CSCptr, double *dCSCval, int *dCSCrow, int *dCSCptr, int N, int NNZ);
+    void CSR2CSC(double *dCSRval, int *dCSRcol, int *dCSRptr, double *CSCval, int *CSCrow, int *CSCptr, double *dCSCval, int *dCSCrow, int *dCSCptr, unsigned long int N, unsigned long int NNZ);
 
-    void CSR2CSC(double *CSRval, int *CSRcol, int *CSRptr, double *CSCval, int *CSCrow, int *CSCptr, int N, int NNZ);
+    void CSR2CSC(double *CSRval, int *CSRcol, int *CSRptr, double *CSCval, int *CSCrow, int *CSCptr, unsigned long int N, unsigned long int NNZ);
 
-    void MtxVec_mult(double *in, double *out, int size, double *val, int *col, int *ptr);
+    void MtxVec_mult(double *in, double *out, unsigned long int size, double *val, int *col, int *ptr);
 
     void MtxVec_mult(double *in, double *out, double *val, int *col, int *ptr);
 
-    void MtxVec_mult(double *in, int inindex, int insize, double *out, int outindex, int outsize, double *val, int *col, int *ptr);
+    void MtxVec_mult(double *in, unsigned long int inindex, unsigned long int insize, double *out, unsigned long int outindex, unsigned long int outsize, double *val, int *col, int *ptr);
 
-    void MtxVec_mult(double *in, int inindex, int insize, double *out, double *val, int *col, int *ptr);
+    void MtxVec_mult(double *in, unsigned long int inindex, unsigned long int insize, double *out, double *val, int *col, int *ptr);
 
-    double dot(double *in1, double *in2, int size);
+    double dot(double *in1, double *in2, unsigned long int size);
 
     double dot(double *in1, double *in2);
 
-    double dot(double *in1, int in1index, int in1size, double *in2, int in2index, int in2size);
+    double dot(double *in1, unsigned long int in1index, unsigned long int in1size, double *in2, unsigned long int in2index, unsigned long int in2size);
 
-    double dot(double *in1, double *in2, int in2index, int in2size);
+    double dot(double *in1, double *in2, unsigned long int in2index, unsigned long int in2size);
 
-    double dot(double *in1, int in1index, int in1size, double *in2);
+    double dot(double *in1, unsigned long int in1index, unsigned long int in1size, double *in2);
 
     void Kskip_cg_bicg_base(double *Ar, double *Ap, double *rvec, double *pvec, const int kskip, double *val, int *col, int *ptr);
 
@@ -83,7 +83,7 @@ class cuda {
 
     void Kskip_bicg_innerProduce2(double *theta, double *eta, double *rho, double *phi, double *Ar, double *Ap, double *r_vec, double *p_vec, int kskip, double *val, int *col, int *ptr);
 
-    void dot_gmres(double *wvec, double *vmtx, double *hmtx, int k, int N);
+    void dot_gmres(double *wvec, double *vmtx, double *hmtx, int k, unsigned long int N);
 };
 #endif //CUDAFUNCTION_HPP_INCLUDED__
